+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Listing BIM360 users
+- Searching BIM360 users using filters
 
 ## [8.0.1] - 2019-11-04